requirements.txt
python-dotenv
requests
playwright
trafilatura